add error banner
