+++ frontend/src/plugins/impl/data-frames/DataFramePlugin.tsx
@@ -13,17 +13,9 @@ import { useAsyncData } from "@/hooks/useAsyncData";
 import { LoadingDataTableComponent } from "../DataTablePlugin";
 import { Functions } from "@/utils/functions";
 import { Arrays } from "@/utils/arrays";
-import { prettyError } from "@/utils/errors";
 import { useState } from "react";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogContent,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from "@/components/ui/alert-dialog";
 import { TooltipProvider } from "@/components/ui/tooltip";
+import { ErrorBanner } from "../common/error-banner";
 
 /**
  * Arguments for a data table
@@ -167,39 +159,5 @@ export const DataFrameComponent = ({
         selection={null}
       />
     </div>
-  );
-};
-
-const ErrorBanner = ({ error }: { error: Error }) => {
-  const [open, setOpen] = useState(false);
-
-  if (!error) {
-    return null;
-  }
-
-  const message = prettyError(error);
-
-  return (
-    <>
-      <div
-        className="text-error border-[var(--red-6)] bg-[var(--red-2)] text-sm p-2 border cursor-pointer hover:bg-[var(--red-3)]"
-        onClick={() => setOpen(true)}
-      >
-        <span className="line-clamp-4">{message}</span>
-      </div>
-      <AlertDialog open={open} onOpenChange={setOpen}>
-        <AlertDialogContent className="max-w-[80%]">
-          <AlertDialogHeader>
-            <AlertDialogTitle className="text-error">Error</AlertDialogTitle>
-          </AlertDialogHeader>
-          <div className="text-error text-sm p-2 font-mono">{message}</div>
-          <AlertDialogFooter>
-            <AlertDialogAction autoFocus={true} onClick={() => setOpen(false)}>
-              Ok
-            </AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
-    </>
   );
 };
